requirements.txt
colorama
mock
matplotlib
numpy
pyyaml
pytest
pytest-html
pytest-mock
pytest-datadir-ng
pytest-pythonpath
pytest-timeout
requests
PyMySQL
pyserial

